+++ AdMob/GoogleAdMobAdapterConfiguration.h
@@ -12,6 +12,7 @@
 #import <MoPubSDKFramework/MoPub.h>
 #else
 #import "MPBaseAdapterConfiguration.h"
+#import "MoPub.h"
 #endif
 
 NS_ASSUME_NONNULL_BEGIN
@@ -28,7 +29,6 @@ NS_ASSUME_NONNULL_BEGIN
  @param parameters Ad response parameters
  */
 + (void)updateInitializationParameters:(NSDictionary *)parameters;
-+ (void)setNpaString:(NSString *)string;
 
 // MPAdapterConfiguration
 @property (nonatomic, copy, readonly) NSString * adapterVersion;