+++ AdMob/GoogleAdMobAdapterConfiguration.m
@@ -12,16 +12,11 @@
     #import "MPLogging.h"
 #endif
 
-@interface GoogleAdMobAdapterConfiguration()
-@property (class, nonatomic, copy, readwrite) NSString * npaString;
-@end
-
 // Initialization configuration keys
 static NSString * const kAdMobApplicationIdKey = @"appid";
 
 // Errors
 static NSString * const kAdapterErrorDomain = @"com.mopub.mopub-ios-sdk.mopub-admob-adapters";
-static NSString * gNpaString = nil;
 
 typedef NS_ENUM(NSInteger, AdMobAdapterErrorCode) {
     AdMobAdapterErrorCodeMissingAppId,
@@ -45,7 +40,7 @@ typedef NS_ENUM(NSInteger, AdMobAdapterErrorCode) {
 #pragma mark - MPAdapterConfiguration
 
 - (NSString *)adapterVersion {
-    return @"7.53.1.0";
+    return @"7.55.0.0";
 }
 
 - (NSString *)biddingToken {
@@ -57,15 +52,11 @@ typedef NS_ENUM(NSInteger, AdMobAdapterErrorCode) {
 }
 
 - (NSString *)networkSdkVersion {
-    return @"7.53.1";
+    return @"7.55.0";
 }
 
 - (void)initializeNetworkWithConfiguration:(NSDictionary<NSString *, id> *)configuration
                                   complete:(void(^)(NSError *))complete {
-    
-    NSString *npaValue = configuration[@"npa"];
-
-    GoogleAdMobAdapterConfiguration.npaString = npaValue;
 
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
@@ -80,14 +71,10 @@ typedef NS_ENUM(NSInteger, AdMobAdapterErrorCode) {
     });
 }
 
+// MoPub collects GDPR consent on behalf of Google
 + (NSString *)npaString
 {
-    return gNpaString;
-}
-
-+ (void)setNpaString:(NSString *)string
-{
-    gNpaString = string;
+    return !MoPub.sharedInstance.canCollectPersonalInfo ? @"1" : @"";
 }
 
 @end