+++ IronSource/IronSourceUtils.h
@@ -2,8 +2,8 @@
 //  IronSourceUtils.h
 //
 #import <Foundation/Foundation.h>
-#if __has_include(<MoPubSDK/MoPub.h>)
-#import <MoPubSDK/MoPub.h>
+#if __has_include(<MoPub/MoPub.h>)
+#import <MoPub/MoPub.h>
 #elif __has_include(<MoPubSDKFramework/MoPub.h>)
 #import <MoPubSDKFramework/MoPub.h>
 #else