+++ Mintegral/MintegralRewardedVideoCustomEvent.h
@@ -1,0 +1,9 @@
+#if __has_include(<MoPubSDKFramework/MoPub.h>)
+    #import <MoPubSDKFramework/MoPub.h>
+#else
+    #import "MPRewardedVideoCustomEvent.h"
+#endif
+
+@interface MintegralRewardedVideoCustomEvent : MPRewardedVideoCustomEvent
+
+@end