+++ AppLovin/CHANGELOG.md
@@ -1,4 +1,10 @@
 ## Changelog
+   * 6.11.3.0
+     * This version of the adapters has been certified with AppLovin SDK 6.11.3.
+
+   * 6.11.1.0
+     * This version of the adapters has been certified with AppLovin SDK 6.11.1.
+
    * 6.11.0.0
      * This version of the adapters has been certified with AppLovin SDK 6.11.0.